+++ app/src/main/java/com/example/leoberteck/cardcrud/repository/interfaces/ITypeRepository.java
@@ -6,5 +6,5 @@ import com.example.leoberteck.cardcrud.entity.Type;
  * Created by Trovata on 22/07/2017.
  */
 
-public interface ITypeRepository extends IAbstractRepository<Type, Integer> {
+public interface ITypeRepository extends IAbstractRepository<Type, Long> {
 }
